requirements.txt
anyio==3.4.0
certifi==2021.10.8
charset-normalizer==2.0.8
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.12.0
hyperframe==6.0.1
idna==3.3
priority==2.0.0
psycopg2==2.9.2
pydantic==1.8.2
python-dotenv==0.19.2
requests==2.26.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
wsproto==1.0.0
